improved formatting

+++ src/use_case/games/war/war_logic/WarSurrenderInteractor.java
@@ -11,27 +11,35 @@ import use_case.games.GameDataAccessInterface;
  * CardsAPIInterface, GameDataAccessInterface, and HistoryDataAccessInterface. It communicates the results
  * to the WarSurrenderOutputBoundary for presentation.
  */
-public class WarSurrenderInteractor implements WarSurrenderInputBoundary{
+public class WarSurrenderInteractor implements WarSurrenderInputBoundary {
 
-    /** Interface for drawing cards from an external API. */
+    /**
+     * Interface for drawing cards from an external API.
+     */
     final CardsAPIInterface cardsAPI;
 
-    /** Interface for accessing game-related data. */
+    /**
+     * Interface for accessing game-related data.
+     */
     final GameDataAccessInterface dataAccess;
 
-    /** Interface for accessing historical game data. */
+    /**
+     * Interface for accessing historical game data.
+     */
     final HistoryDataAccessInterface historyDAO;
 
-    /** Presenter for the "Surrender" action. */
+    /**
+     * Presenter for the "Surrender" action.
+     */
     final WarSurrenderOutputBoundary warSurrenderPresenter;
 
     /**
      * Constructs a new WarSurrenderInteractor with the specified dependencies.
      *
-     * @param cardsAPI                Interface for drawing cards from an external API.
-     * @param dataAccess              Interface for accessing game-related data.
-     * @param historyDAO              Interface for accessing historical game data.
-     * @param warSurrenderPresenter   Presenter for the "Surrender" action.
+     * @param cardsAPI              Interface for drawing cards from an external API.
+     * @param dataAccess            Interface for accessing game-related data.
+     * @param historyDAO            Interface for accessing historical game data.
+     * @param warSurrenderPresenter Presenter for the "Surrender" action.
      */
     public WarSurrenderInteractor(CardsAPIInterface cardsAPI,
                                   GameDataAccessInterface dataAccess,
@@ -55,7 +63,7 @@ public class WarSurrenderInteractor implements WarSurrenderInputBoundary{
         String username = warInputGameData.getUser();
 
         dataAccess.editFund(username, bet / 2);
-        historyDAO.addPayout(username, "war", - (bet / 2));
+        historyDAO.addPayout(username, "war", -((double) bet / 2));
         warSurrenderPresenter.prepareSurrenderView(new WarOutputGameData(game));
     }
 }
